Always expose the quit menu option
* always expose the quit menu option

* rename variable to reflect that it doesn't always deletes data

+++ app/src/main/java/net/waterfox/android/home/HomeMenuBuilder.kt
@@ -141,17 +141,21 @@ class HomeMenuBuilder(
                 )
             }
             HomeMenu.Item.Quit -> {
-                // We need to show the snackbar while the browsing data is deleting (if "Delete
-                // browsing data on quit" is activated). After the deletion is over, the snackbar
-                // is dismissed.
-                deleteAndQuit(
-                    activity = homeActivity,
-                    coroutineScope = lifecycleOwner.lifecycleScope,
-                    snackbar = WaterfoxSnackbar.make(
-                        view = view,
-                        isDisplayedWithBrowserToolbar = false
+                if (context.settings().shouldDeleteBrowsingDataOnQuit) {
+                    // We need to show the snackbar while the browsing data is deleting (if "Delete
+                    // browsing data on quit" is activated). After the deletion is over, the snackbar
+                    // is dismissed.
+                    deleteAndQuit(
+                        activity = homeActivity,
+                        coroutineScope = lifecycleOwner.lifecycleScope,
+                        snackbar = WaterfoxSnackbar.make(
+                            view = view,
+                            isDisplayedWithBrowserToolbar = false,
+                        ),
                     )
-                )
+                } else {
+                    homeActivity.finishAndRemoveTask()
+                }
             }
             HomeMenu.Item.ReconnectSync -> {
                 navController.nav(
